pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.16",
    "langchain-tavily>=0.2.18",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
